+++ signup/src/main/AndroidManifest.xml
@@ -2,28 +2,24 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
           package="nl.markv.brocast.brocast.signup">
 
-    <application>
-        <activity
-                android:name=".SignupActivity"
-                android:label="@string/app_name"
-                android:theme="@style/AppTheme.NoActionBar">
-            <intent-filter android:order="1">
-                <action android:name="android.intent.action.VIEW"/>
+    <uses-permission android:name="android.permission.INTERNET" />
 
-                <category android:name="android.intent.category.BROWSABLE"/>
-                <category android:name="android.intent.category.DEFAULT"/>
-
-                <data
-                        android:host="brocast.markv.nl"
-                        android:pathPattern="/.*"
-                        android:scheme="https"/>
-            </intent-filter>
+    <application
+        android:allowBackup="true"
+        android:icon="@mipmap/ic_launcher"
+        android:label="@string/app_name"
+        android:roundIcon="@mipmap/ic_launcher_round"
+        android:supportsRtl="true"
+        android:theme="@style/AppTheme">
+        <activity android:name=".PhoneAuthActivity">
             <intent-filter>
-                <action android:name="android.intent.action.MAIN"/>
+                <action android:name="android.intent.action.MAIN" />
 
-                <category android:name="android.intent.category.LAUNCHER"/>
+                <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+        <!--<activity android:name=".PhoneAuthWithSdk" />-->
+        <activity android:name=".SignedInActivity"></activity>
     </application>
 
 </manifest>
