+++ signup/src/main/res/layout/activity_signed_in.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:padding="16dp"
-    tools:context=".SignupActivity">
+    tools:context=".SignedInActivity">
 
     <TextView
         android:layout_width="wrap_content"
